backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
python-telegram-bot==21.8
pydantic==2.7.1
